+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## Next
+
+## Added
+- Client can fetch an idle stream using `Ace.HTTP2.Client.stream/1`.
+- Send generic data over a stream using `Ace.HTTP2.Client.send/1`.
+- Forward stream resets with reason to worker processes.
+
 ## [0.11.1](https://github.com/CrowdHailer/Ace/tree/0.11.1) - 2017-08-03
 
 ## Added